+++ src/lib/Components/Inbox/ActiveBids/ActiveBid.tsx
@@ -2,14 +2,18 @@ import OpaqueImageView from "lib/Components/OpaqueImageView"
 import colors from "lib/data/colors"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import React from "react"
-import { TouchableWithoutFeedback } from "react-native"
+import { Dimensions, TouchableWithoutFeedback } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
 import styled from "styled-components/native"
 import { BodyText, MetadataText } from "../Typography"
 
+const isPad = Dimensions.get("window").width > 700
+
 const Container = styled.View`
-  margin: 17px 20px 0;
-  height: 80px;
+  height: 120px;
+  margin-left: 20px;
+  margin-right: 20px;
+  ${isPad ? "align-self: center; width: 708;" : ""};
 `
 
 const Content = styled.View`
@@ -19,13 +23,13 @@ const Content = styled.View`
 `
 
 const ImageView = styled(OpaqueImageView)`
-  width: 58px;
-  height: 58px;
+  width: 80px;
+  height: 80px;
   border-radius: 4;
 `
 
 const MetadataContainer = styled.View`
-  margin: 0 15px;
+  margin: 0 10px;
   flex: 1;
   flex-direction: column;
   justify-content: center;
@@ -35,10 +39,10 @@ const Separator = styled.View`
   height: 1;
   width: 100%;
   background-color: ${colors["gray-regular"]};
-  margin-top: 17px;
 `
 
 const StatusLabel = styled(MetadataText)`
+  margin-bottom: 3px;
   color: ${(props: { status: BidStatus }) => {
     switch (props.status) {
       case "winning":
@@ -129,6 +133,9 @@ class ActiveBid extends React.Component<RelayProps, State> {
           <Content>
             <ImageView imageURL={imageURL} />
             <MetadataContainer>
+              <StatusLabel status={this.state.status}>
+                {this.statusLabel}
+              </StatusLabel>
               <BodyText>
                 {headline}
               </BodyText>
@@ -136,9 +143,6 @@ class ActiveBid extends React.Component<RelayProps, State> {
                 {subtitle}
               </BodyText>
             </MetadataContainer>
-            <StatusLabel status={this.state.status}>
-              {this.statusLabel}
-            </StatusLabel>
           </Content>
           <Separator />
         </Container>