+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 -   Change inquiry status bar background to white - maxim
 -   Move pagination from home/sale index to LotsByfollowedArtists, misc QA - chris
 -   Only load a single Works For You page at a time - alloy
+-   QA on Inbox and Active Bids spacing - maxim
+-   iPad support for Active Bids - maxim
 -   Get rid of image view flag that would skip on-the-fly resizing - alloy
 
 ### 1.4.0-beta.9